requirements.txt
numpy
matplotlib
Pillow
tk
opencv-python
scikit-learn